test/ru/job4j/calculator/PointTest.java
package ru.job4j.calculator;

import org.junit.Assert;
import org.junit.Test;

public class PointTest {

    @Test
    public void when00to20then2() {
        int x1 = 0;
        int y1 = 0;
        int x2 = 2;
        int y2 = 0;
        int expected = 2;
        double out = ru.job4j.condition.Point.distance(x1,y1,x2,y2);
        Assert.assertEquals(out,expected,0.01);
    }
    @Test
    public void when21to31then1() {
        int x1 = 2;
        int y1 = 1;
        int x2 = 3;
        int y2 = 1;
        int expected = 1;
        double out = ru.job4j.condition.Point.distance(x1,y1,x2,y2);
        Assert.assertEquals(out,expected,0.01);
    }
    @Test
    public void when22to65then5() {
        int x1 = 2;
        int y1 = 2;
        int x2 = 6;
        int y2 = 5;
        int expected = 5;
        double out = ru.job4j.condition.Point.distance(x1,y1,x2,y2);
        Assert.assertEquals(expected,out,0.01);
    }
}